Removed ViewsController functions from DefaultController and created a Handler to host the functions that both Controllers use
